requirements.txt
streamlit>=1.30.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
pymilvus>=2.3.0
litellm>=1.0.0
PyMuPDF>=1.23.0
tenacity>=8.2.0
pytest>=8.0.0
langchain-core>=0.1.27
langchain-community>=0.0.24
langchain-openai>=0.0.5
tqdm>=4.66.0
langfuse